Move tests for VM from main.rs to test crate.
