+++ src/main.rs
@@ -128,10 +128,7 @@ fn repl(is_profile: bool, is_trace: bool) {
 
                     match global_context {
                         Some(ref mut context) => {
-                            context.append_from_function_info(
-                                &mut vm.factory,
-                                &global_info,
-                            );
+                            context.append_from_function_info(&mut vm.factory, &global_info);
                             context.module_func_id = global_info.module_func_id;
                             context.func_id = global_info.func_id;
                             context.func_ref = global_info;
@@ -171,197 +168,5 @@ fn repl(is_profile: bool, is_trace: bool) {
                 }
             }
         }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use rapidus::test::{assert_file, execute_script, runtime_error, test_code, test_file};
-
-    #[test]
-    fn vm_test() {
-        execute_script("for(var i = 0; i < 4; i++){ i }".to_string());
-        // TODO: Following tests should run. Fix them ASAP.
-
-        // assert_file("letconst".to_string());
-        // assert_file("nested_block".to_string());
-        // assert_file("nested_block2".to_string());
-        // test_code("'true'*3".to_string(), "'truetruetrue'".to_string());
-        // test_code("(100).toString(15)".to_string(), "'6a'".to_string());
-        // test_file(
-        //     "label".to_string(),
-        //     "[ 0, 0, 0, 1, 0, 2, 1, 0, 2, 0, 3, 0, 3, 1, 4, 1, 4, 2, 0 ]".to_string(),
-        // );
-
-        //test_file(
-        //    "qsort".to_string(),
-        //    "[ 0, 0, 1, 3, 5, 7, 7, 10, 11, 12, 14, 14, 16, 17, 19 ]".to_string(),
-        //);
-        // test_file("arguments1".to_string(), "[[1,2,3,4,4],[1,2,[3,4]],[5,6,7,undefined,3],[5,6,[7]],[8,9,undefined,undefined,2],[8,9,undefined],[10,undefined,undefined,undefined,1],[10,undefined,undefined]]".to_string());
-        // test_file(
-        //     "arguments2".to_string(),
-        //     "[10,15,20,25,15,10,'OK',20,25,'OK',10,'NG',20,25,'NG']".to_string(),
-        // );
-    }
-
-    #[test]
-    fn string_test1() {
-        test_code("'死して屍拾う者なし'[4]", "'拾'");
-    }
-
-    #[test]
-    fn string_test2() {
-        test_code("'死して屍拾う者なし'.length", "9");
-    }
-    #[test]
-    fn operator_test() {
-        test_code("+(5>3)+60%7+(3>=5)+!!5+(-6)", "0");
-    }
-
-    #[test]
-    fn operator_test2() {
-        assert_file("operator");
-    }
-
-    #[test]
-    fn this_test() {
-        test_file("this", "[1,101,124]");
-    }
-
-    #[test]
-    fn prototype_test() {
-        assert_file("prototypes");
-    }
-
-    #[test]
-    fn accessor_property() {
-        assert_file("accessor_property");
-    }
-
-    #[test]
-    fn trinity() {
-        assert_file("trinity");
-    }
-
-    #[test]
-    fn closure() {
-        assert_file("closure");
-    }
-
-    #[test]
-    fn trycatch() {
-        assert_file("trycatch");
-    }
-
-    #[test]
-    fn r#typeof() {
-        assert_file("typeof");
-    }
-
-    #[test]
-    fn exotic_cmp() {
-        assert_file("exotic_cmp")
-    }
-
-    #[test]
-    fn r#while() {
-        assert_file("while")
-    }
-
-    #[test]
-    fn r#for() {
-        assert_file("for")
-    }
-
-    #[test]
-    fn r#if() {
-        assert_file("if")
-    }
-
-    #[test]
-    fn arrow_function() {
-        test_code("let f = (x) => { return x * x }; f(5)", "25");
-        test_code("let f = x => { return x * x }; f(6)", "36");
-    }
-
-    #[test]
-    fn symbol() {
-        assert_file("symbol")
-    }
-
-    #[test]
-    fn array() {
-        assert_file("array")
-    }
-
-    #[test]
-    fn env() {
-        assert_file("env")
-    }
-
-    #[test]
-    fn new_call_member() {
-        assert_file("new_call_member")
-    }
-
-    #[test]
-    fn assert() {
-        assert_file("assert")
-    }
-
-    #[test]
-    fn fibo() {
-        assert_file("fibo")
-    }
-
-    #[test]
-    fn fact() {
-        assert_file("fact")
-    }
-
-    #[test]
-    fn test_module() {
-        assert_file("test_module_caller")
-    }
-
-    #[test]
-    fn function_methods() {
-        assert_file("function_methods")
-    }
-
-    #[test]
-    fn string_methods() {
-        assert_file("string_methods")
-    }
-
-    #[test]
-    fn runtime_error1() {
-        runtime_error("let a = {}; a.b.c");
-    }
-
-    #[test]
-    fn runtime_error2() {
-        runtime_error("let a = {}; a.b.c = 5");
-    }
-
-    #[test]
-    fn runtime_error3() {
-        runtime_error("let a = {}; a(5)");
-    }
-
-    #[test]
-    fn runtime_error4() {
-        runtime_error("let a = {}; a.b(5)");
-    }
-
-    #[test]
-    fn runtime_error5() {
-        runtime_error("let a = {}; a(5)");
-    }
-
-    use test::Bencher;
-    #[bench]
-    fn bench_fibo(b: &mut Bencher) {
-        b.iter(|| assert_file("fibo"));
     }
 }
