+++ tests/test/accessor_property.js
