+++ tests/test/arguments2.js
