+++ tests/test/closure.js
