+++ tests/test/exotic_cmp.js
