+++ tests/test/for.js
