+++ tests/test/function_methods.js
