+++ tests/test/label.js
