+++ tests/test/operator.js
