+++ tests/test/string_methods.js
