+++ tests/test/symbol.js
