+++ tests/test/test_module_caller.js
@@ -1,3 +1,3 @@
-let mod = require('./test/test_module.js')
+let mod = require('./tests/test/test_module.js')
 mod.countup(5)
 if (mod.count() !== 15) throw new Error()
