+++ tests/test/trinity.js
