+++ tests/test/trycatch.js
