+++ tests/test/typeof.js
